Moved RapidsShuffleExceptions.scala to a rapids subdir
